+++ dbt/adapters/databricks/relation_configs/streaming_table.py
@@ -8,6 +8,9 @@ from dbt.adapters.databricks.relation_configs.base import (
 from dbt.adapters.databricks.relation_configs.comment import (
     CommentProcessor,
 )
+from dbt.adapters.databricks.relation_configs.liquid_clustering import (
+    LiquidClusteringProcessor,
+)
 from dbt.adapters.databricks.relation_configs.partitioning import (
     PartitionedByProcessor,
 )
@@ -22,6 +25,7 @@ from dbt.adapters.databricks.relation_configs.tblproperties import (
 class StreamingTableConfig(DatabricksRelationConfigBase):
     config_components = [
         PartitionedByProcessor,
+        LiquidClusteringProcessor,
         CommentProcessor,
         TblPropertiesProcessor,
         RefreshProcessor,